Fix `_rg_internals.update_stream_last_read_id` command definition.
The PR fixes `_rg_internals.update_stream_last_read_id` command definition with respect
to key spec. Though the command is internal, it is send to replica (and replica of) and
it is required that the command definition will be done correctly.

In addition, the PR makes sure that all internal command will only run if came from primary.
Tests was added to verify it.
