+++ redisgears_core/src/function_load_command.rs
@@ -5,13 +5,12 @@
  */
 
 use redis_module::{
-    Context, ContextFlags, NextArg, RedisError, RedisResult, RedisString, RedisValue,
-    ThreadSafeContext,
+    Context, NextArg, RedisError, RedisResult, RedisString, RedisValue, ThreadSafeContext,
 };
 use redisgears_plugin_api::redisgears_plugin_api::GearsApiError;
 
 use crate::compiled_library_api::CompiledLibraryAPI;
-use crate::{get_backend, get_globals, verify_name, Deserialize, Serialize};
+use crate::{get_backend, verify_name, Deserialize, Serialize};
 
 use crate::{
     get_libraries, GearsLibrary, GearsLibraryCtx, GearsLibraryMetaData, GearsLoadLibraryCtx,
@@ -331,32 +330,10 @@ pub(crate) fn function_load_command(
     Ok(RedisValue::NoReply)
 }
 
-/// Verify that it is OK to perform an internal command.
-/// Internal command should only run if it came from replication stream
-/// or from AOF loading. In other words, the command did not came directly from a user.
-fn verify_internal_command(ctx: &Context) -> Result<(), RedisError> {
-    let flags = ctx.get_flags();
-    let globals = get_globals();
-    if !flags.contains(ContextFlags::LOADING)
-        && !(flags.contains(ContextFlags::REPLICATED) || globals.db_policy.is_pseudo_slave())
-    {
-        // Internal commands should either be loaded from AOF ([`ContextFlags::LOADING`])
-        // or sent over the replication stream([`ContextFlags::REPLICATED`]).
-        // Another special option is if the instance is pseudo slave (replica of) which is treated as if the command was replicated from primary.
-        // If none of those cases holds we will return an error.
-        return Err(RedisError::Str(
-            "Internal command should only be sent from primary or loaded from AOF",
-        ));
-    }
-    Ok(())
-}
-
 pub(crate) fn function_load_on_replica(
     ctx: &Context,
     args: Skip<IntoIter<redis_module::RedisString>>,
 ) -> RedisResult {
-    verify_internal_command(ctx)?;
-
     let args = get_args_values(args)?;
     if args.user.is_none() {
         return Err(RedisError::Str("User was not provided by primary"));
